Movie-Recommendation-BE/pyproject.toml
[tool.poetry]
name = "movie-recommendation-be"
version = "0.1.0"
description = "Movie Recommendation API"
authors = ["BillyMwangiDev"]
readme = "README.md"
packages = [{include = "config"}, {include = "apps"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "5.1.5"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
django-filter = "23.5"
psycopg2-binary = "2.9.9"
dj-database-url = "2.1.0"
djangorestframework-simplejwt = "5.3.1"
drf-yasg = "1.21.7"
requests = "2.32.5"
django-redis = "5.4.0"
redis = "5.0.1"
celery = "5.3.4"
django-celery-beat = "2.7.0"
django-celery-results = "2.5.1"
kombu = "5.3.4"
graphene-django = "3.2.0"
django-graphql-jwt = "0.4.0"
graphene-django-optimizer = "0.10.0"
gunicorn = "21.2.0"
whitenoise = "6.4.0"
python-decouple = "3.8"
python-dotenv = "1.0.0"
sentry-sdk = "1.39.2"
django-ratelimit = "4.1.0"
django-axes = "6.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-django = "4.7.0"
pytest-cov = "4.1.0"
factory-boy = "3.3.0"
faker = "21.0.0"
black = "23.12.1"
flake8 = "7.0.0"
isort = "5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


